Add condition if os.file cluster assignment exists

+++ prueba_COAD/scripts/get_clusters_ctpsingle.py
@@ -16,22 +16,28 @@ cluster_size_file = os.path.join(cwd,'data','ctpsingle','clusters_size_per_sampl
 fi=open(cluster_size_file,'w')
 fi.write("sample\tcluster\tsize\n")
 
-for sample in samples[0:2]:
+for sample in samples:
     
     input_path=os.path.join(cwd,'data','ctpsingle',sample,(sample + "_ctpsingle_cluster_assignments.txt"))
-    clusters_file=pd.read_csv(input_path, sep='\t')
+    if os.path.exists(input_path):
+        clusters_file=pd.read_csv(input_path, sep='\t')
     
-    #Add lines with number of clusters per sample
-    num_clusters=clusters_file['mostLikely'].nunique()
-    print(sample+"\t"+str(num_clusters), file=fh)
+        #Add lines with number of clusters per sample
+        num_clusters=clusters_file['mostLikely'].nunique()
+        print(sample+"\t"+str(num_clusters), file=fh)
 
-    #Add lines with sizes of clusters from each sample
-    cluster_size=clusters_file['mostLikely'].value_counts()
-    cluster_size_df=cluster_size.to_frame()
-    cluster_size_df.columns = ['size']
-    cluster_size_df.insert(0, "cluster", cluster_size.index) #Add column with cluster id (index)
-    cluster_size_df.insert(0, "sample", sample) #Add colum with sample code
-    cluster_size_df.to_csv(fi,sep='\t',header=False, index =False)
+        #Add lines with sizes of clusters from each samp    le
+        cluster_size=clusters_file['mostLikely'].value_counts()
+        cluster_size_df=cluster_size.to_frame()
+        cluster_size_df.columns = ['size']
+        cluster_size_df.insert(0, "cluster", cluster_size.index) #Add column with cluster id (index)
+        cluster_size_df.insert(0, "sample", sample) #Add colum with sample code
+        cluster_size_df.to_csv(fi,sep='\t',header=False, index =False)
+    else:
+        continue
+
 fh.close()
 fi.close()
+
+print("Work finished, good bye!")
